+++ packages/flutter_tools/lib/runner.dart
@@ -17,12 +17,12 @@ import 'src/base/logger.dart';
 import 'src/base/process.dart';
 import 'src/base/utils.dart';
 import 'src/context_runner.dart';
-import 'src/crash_reporting.dart';
 import 'src/doctor.dart';
 import 'src/globals.dart';
+import 'src/reporting/crash_reporting.dart';
+import 'src/reporting/usage.dart';
 import 'src/runner/flutter_command.dart';
 import 'src/runner/flutter_command_runner.dart';
-import 'src/usage.dart';
 import 'src/version.dart';
 
 /// Runs the Flutter tool with support for the specified list of [commands].
@@ -109,18 +109,19 @@ Future<int> _handleToolError(
       stderr.writeln(stackTrace.toString());
       return _exit(1);
     } else {
-      flutterUsage.sendException(error, stackTrace);
+      // Report to both [Usage] and [CrashReportSender].
+      flutterUsage.sendException(error);
+      await CrashReportSender.instance.sendReport(
+        error: error,
+        stackTrace: stackTrace,
+        getFlutterVersion: getFlutterVersion,
+      );
 
       if (error is String)
         stderr.writeln('Oops; flutter has exited unexpectedly: "$error".');
       else
         stderr.writeln('Oops; flutter has exited unexpectedly.');
 
-      await CrashReportSender.instance.sendReport(
-        error: error,
-        stackTrace: stackTrace,
-        getFlutterVersion: getFlutterVersion,
-      );
       try {
         final File file = await _createLocalCrashReport(args, error, stackTrace);
         stderr.writeln(
