+++ packages/flutter_tools/lib/src/commands/config.dart
@@ -8,8 +8,8 @@ import '../android/android_sdk.dart';
 import '../android/android_studio.dart';
 import '../convert.dart';
 import '../globals.dart';
+import '../reporting/usage.dart';
 import '../runner/flutter_command.dart';
-import '../usage.dart';
 
 class ConfigCommand extends FlutterCommand {
   ConfigCommand({ bool verboseHelp = false }) {
