+++ packages/flutter_tools/lib/src/context_runner.dart
@@ -41,8 +41,8 @@ import 'macos/cocoapods_validator.dart';
 import 'macos/macos_workflow.dart';
 import 'macos/xcode.dart';
 import 'macos/xcode_validator.dart';
+import 'reporting/usage.dart';
 import 'run_hot.dart';
-import 'usage.dart';
 import 'version.dart';
 import 'web/chrome.dart';
 import 'web/workflow.dart';
