+++ packages/flutter_tools/lib/src/reporting/disabled_usage.dart
@@ -35,7 +35,7 @@ class DisabledUsage implements Usage {
   void sendTiming(String category, String variableName, Duration duration, { String label }) { }
 
   @override
-  void sendException(dynamic exception, StackTrace trace) { }
+  void sendException(dynamic exception) { }
 
   @override
   Stream<Map<String, dynamic>> get onSend => null;
