+++ packages/flutter_tools/lib/src/reporting/usage.dart
@@ -7,13 +7,13 @@ import 'dart:async';
 import 'package:meta/meta.dart';
 import 'package:usage/usage_io.dart';
 
-import 'base/context.dart';
-import 'base/file_system.dart';
-import 'base/os.dart';
-import 'base/platform.dart';
-import 'base/utils.dart';
-import 'globals.dart';
-import 'version.dart';
+import '../base/context.dart';
+import '../base/file_system.dart';
+import '../base/os.dart';
+import '../base/platform.dart';
+import '../base/utils.dart';
+import '../globals.dart';
+import '../version.dart';
 
 const String _kFlutterUA = 'UA-67589403-6';
 
@@ -152,9 +152,9 @@ class Usage {
     }
   }
 
-  void sendException(dynamic exception, StackTrace trace) {
+  void sendException(dynamic exception) {
     if (!suppressAnalytics)
-      _analytics.sendException('${exception.runtimeType}\n${sanitizeStacktrace(trace)}');
+      _analytics.sendException(exception.runtimeType.toString());
   }
 
   /// Fires whenever analytics data is sent over the network.
