+++ packages/flutter_tools/lib/src/runner/flutter_command_runner.dart
@@ -30,8 +30,8 @@ import '../convert.dart';
 import '../dart/package_map.dart';
 import '../device.dart';
 import '../globals.dart';
+import '../reporting/usage.dart';
 import '../tester/flutter_tester.dart';
-import '../usage.dart';
 import '../version.dart';
 import '../vmservice.dart';
