+++ packages/flutter_tools/lib/src/web/compile.dart
@@ -13,7 +13,7 @@ import '../build_info.dart';
 import '../bundle.dart';
 import '../globals.dart';
 import '../project.dart';
-import '../usage.dart';
+import '../reporting/usage.dart';
 
 /// The [WebCompilationProxy] instance.
 WebCompilationProxy get webCompilationProxy => context.get<WebCompilationProxy>();
