+++ packages/flutter_tools/test/general.shard/crash_reporting_test.dart
@@ -17,7 +17,7 @@ import 'package:flutter_tools/src/base/context.dart';
 import 'package:flutter_tools/src/base/io.dart';
 import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/cache.dart';
-import 'package:flutter_tools/src/crash_reporting.dart';
+import 'package:flutter_tools/src/reporting/crash_reporting.dart';
 import 'package:flutter_tools/src/runner/flutter_command.dart';
 import 'package:pedantic/pedantic.dart';
