+++ packages/flutter_tools/test/src/testbed.dart
@@ -16,7 +16,7 @@ import 'package:flutter_tools/src/base/logger.dart';
 import 'package:flutter_tools/src/base/terminal.dart';
 import 'package:flutter_tools/src/cache.dart';
 import 'package:flutter_tools/src/context_runner.dart';
-import 'package:flutter_tools/src/usage.dart';
+import 'package:flutter_tools/src/reporting/usage.dart';
 import 'package:flutter_tools/src/version.dart';
 
 import 'context.dart';
@@ -164,7 +164,7 @@ class NoOpUsage implements Usage {
   void sendEvent(String category, String parameter,{ Map<String, String> parameters }) {}
 
   @override
-  void sendException(dynamic exception, StackTrace trace) {}
+  void sendException(dynamic exception) {}
 
   @override
   void sendTiming(String category, String variableName, Duration duration, { String label }) {}
